feat: added collapsible feature to recent files list and refactored recent-files-section-heading to chat-info-section-heading
